requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-multipart